Move helper functions to utilities
Having a separate import for helpers enables custom nodes using them.

Refactored the input widgets extension to use the helpers.
